Fixing incorrect assumption that ulong was as big as size_t.
Copied from Perforce
 Change: 178199
 ServerID: perforce.ravenbrook.com

+++ mps/code/eventcnv.c
@@ -413,7 +413,7 @@ static void readLog(EventProc proc)
 
   while (TRUE) { /* loop for each event */
     char *eventFormat;
-    int argCount, i;
+    size_t argCount, i;
     Event event;
     EventCode code;
     Res res;
